Suggested revisions to code organization
Here is a summary of the changes I suggest through this PR:

- File prologue: Include a high-level problem statement and a URL to the actual problem definition.

- General: use `const` where possible. Please start a SIGCPP thread so we can discus this point further

- L9: Space before `<iostream>`

- L17: variable sum is now declared and initialized at the latest point possible

- L19: use qualified name `std::cout` instead of inserting a qualified name earlier with a `using` declaration. The name `cout` is used only once.

Open questions: The algorithm/strategy is unclear. Specifically, variable names `N` and `M` do not communicate their purpose. These variables names need to be revised to match what they represent. Also, a comment is needed to explain the algorithm/strategy.